+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased - [0.2.2]
+
+### Added
+
+- [web-client](./web-client) PATCH added a checkbox component
+
+### Changed
+
+### Fixed
+
 ## [0.2.1]
 
 ### Added